+++ frontend/src/hooks/useKeyDown/constants.ts
@@ -1,0 +1,12 @@
+const KEY_ACTIONS = {
+  vertical: {
+    ArrowDown: 1,
+    ArrowUp: -1,
+  },
+  horizontal: {
+    ArrowRight: 1,
+    ArrowLeft: -1,
+  },
+};
+
+export default KEY_ACTIONS;
